refactor ProductList
- now we don't have two components for the same product tile design
