+++ project-base/storefront/components/Blocks/Product/ProductsList/ProductListItemSkeleton.tsx
@@ -1,6 +1,6 @@
 import Skeleton from 'react-loading-skeleton';
 
-export const ProductItemSkeleton: FC = () => (
+export const ProductListItemSkeleton: FC = () => (
     <div className="flex w-full flex-col border-b border-greyLighter pb-5 lg:py-4 lg:px-5">
         <Skeleton className="mb-24 h-full rounded-none" containerClassName="h-[179px] mb-24 max-w-[160px] w-full" />
         <Skeleton className="mb-4 h-4 w-5/6" containerClassName="w-full" />
