+++ project-base/storefront/components/Blocks/Product/ProductsList/ProductsList.tsx
@@ -1,16 +1,12 @@
-import { ProductItem } from './ProductItem';
 import { Adverts } from 'components/Blocks/Adverts/Adverts';
 import { DEFAULT_PAGE_SIZE } from 'components/Blocks/Pagination/Pagination';
 import { CategoryDetailFragmentApi, ListedProductFragmentApi } from 'graphql/generated';
 import { createEmptyArray } from 'helpers/arrayUtils';
-import { useQueryParams } from 'hooks/useQueryParams';
-import { useWishlist } from 'hooks/useWishlist';
 import React from 'react';
 import { GtmMessageOriginType, GtmProductListNameType } from 'types/gtm/enums';
-import { ProductItemSkeleton } from './ProductItemSkeleton';
+import { ProductListItemSkeleton } from './ProductListItemSkeleton';
 import { CategoryDetailContentMessage } from 'components/Pages/CategoryDetail/CategoryDetailContentMessage';
-import { useComparison } from 'hooks/comparison/useComparison';
-import { ProductComparePopup } from '../ButtonsAction/ProductComparePopup';
+import { ProductsListContent } from './ProductsListContent';
 
 type ProductsListProps = {
     products: ListedProductFragmentApi[] | undefined;
@@ -29,11 +25,6 @@ export const ProductsList: FC<ProductsListProps> = ({
     category,
     gtmMessageOrigin = GtmMessageOriginType.other,
 }) => {
-    const { currentPage } = useQueryParams();
-    const { isPopupCompareOpen, toggleProductInComparison, setIsPopupCompareOpen, isProductInComparison } =
-        useComparison();
-    const { toggleProductInWishlist, isProductInWishlist } = useWishlist();
-
     if (!products?.length && !fetching) {
         return <CategoryDetailContentMessage />;
     }
@@ -45,19 +36,11 @@ export const ProductsList: FC<ProductsListProps> = ({
         >
             {!!products?.length && !fetching ? (
                 <>
-                    {products.map((product, index) => (
-                        <ProductItem
-                            key={product.uuid}
-                            product={product}
-                            listIndex={(currentPage - 1) * DEFAULT_PAGE_SIZE + index}
-                            gtmProductListName={gtmProductListName}
-                            gtmMessageOrigin={gtmMessageOrigin}
-                            isProductInComparison={isProductInComparison(product.uuid)}
-                            toggleProductInComparison={() => toggleProductInComparison(product.uuid)}
-                            isProductInWishlist={isProductInWishlist(product.uuid)}
-                            toggleProductInWishlist={() => toggleProductInWishlist(product.uuid)}
-                        />
-                    ))}
+                    <ProductsListContent
+                        products={products}
+                        gtmProductListName={gtmProductListName}
+                        gtmMessageOrigin={gtmMessageOrigin}
+                    />
 
                     {category && (
                         <Adverts
@@ -67,14 +50,9 @@ export const ProductsList: FC<ProductsListProps> = ({
                             isSingle
                         />
                     )}
-
-                    <ProductComparePopup
-                        isVisible={isPopupCompareOpen}
-                        onCloseCallback={() => setIsPopupCompareOpen(false)}
-                    />
                 </>
             ) : (
-                createEmptyArray(DEFAULT_PAGE_SIZE).map((_, index) => <ProductItemSkeleton key={index} />)
+                createEmptyArray(DEFAULT_PAGE_SIZE).map((_, index) => <ProductListItemSkeleton key={index} />)
             )}
         </div>
     );
